requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
httpx
google-auth
google-auth-oauthlib
google-api-python-client
loguru
python-multipart
psycopg2-binary 
defusedxml==0.7.1
cryptography
